refactor background system to environment system
