+++ packages/engine/src/scene/systems/SceneObjectSystem.tsx
@@ -36,7 +36,7 @@ import { GroupComponent, Object3DWithEntity, startGroupQueryReactor } from '../c
 import { ShadowComponent } from '../components/ShadowComponent'
 import { UpdatableCallback, UpdatableComponent } from '../components/UpdatableComponent'
 import { VisibleComponent } from '../components/VisibleComponent'
-import BackgroundSystem from './BackgroundSystem'
+import EnvironmentSystem from './EnvironmentSystem'
 import FogSystem from './FogSystem'
 import ShadowSystem from './ShadowSystem'
 
@@ -165,12 +165,11 @@ export default async function SceneObjectSystem() {
     await groupReactor.stop()
   }
 
-  const subsystems = [() => Promise.resolve({ default: FogSystem })]
-  if (isClient)
-    subsystems.push(
-      () => Promise.resolve({ default: ShadowSystem }),
-      () => Promise.resolve({ default: BackgroundSystem })
-    )
+  const subsystems = [
+    () => Promise.resolve({ default: FogSystem }),
+    () => Promise.resolve({ default: EnvironmentSystem })
+  ]
+  if (isClient) subsystems.push(() => Promise.resolve({ default: ShadowSystem }))
 
   return {
     execute,
